+++ src/langserver/toolsDocumentSymbolsRequest.ts
@@ -1,10 +1,12 @@
 export class ToolsDocumentSymbolsRequest {
     source : string;
     path : string;
+    includeProperties : boolean;
 
-    constructor(newSourceCode : string, newPath: string) {
+    constructor(newSourceCode : string, newPath: string, newIncludeProperties: boolean) {
         this.source = newSourceCode;
         this.path = newPath;
+        this.includeProperties = newIncludeProperties;
     }
 
 }