+++ src/objectwizards/symbolwizards/alSymbolsBasedQueryWizard.ts
@@ -102,7 +102,7 @@ export class ALSymbolsBasedQueryWizard extends ALSymbolsBasedWizard {
         writer.writeStartNameSourceBlock("dataitem", dataItemName, writer.encodeName(tableSymbol.name));
 
         let fieldList : AZSymbolInformation[] = [];
-        tableSymbol.collectChildSymbols(AZSymbolKind.Field, fieldList);
+        tableSymbol.collectChildSymbols(AZSymbolKind.Field, true, fieldList);
         fieldList.forEach(
             item => {
                 writer.writeNameSourceBlock("column", writer.createName(item.name), writer.encodeName(item.name));