+++ src/symbollibraries/azSymbolInformation.ts
@@ -92,6 +92,8 @@ export class AZSymbolInformation {
         (this.kind == AZSymbolKind.EnumType) ||
         (this.kind == AZSymbolKind.EnumExtensionType) ||
         (this.kind == AZSymbolKind.DotNetPackage) ||
+        (this.kind == AZSymbolKind.ProfileObject) ||
+        (this.kind == AZSymbolKind.PageCustomizationObject) ||
         (this.kind == AZSymbolKind.Interface));
     }
 
@@ -272,15 +274,28 @@ export class AZSymbolInformation {
         return 'undefined';
     }
 
-    public collectChildSymbols(symbolKind: AZSymbolKind, outList : AZSymbolInformation[]) {
+    public collectChildSymbols(symbolKind: AZSymbolKind, includeAllLevels: boolean, outList : AZSymbolInformation[]) {
         if (this.childSymbols) {
             for (let i=0; i<this.childSymbols.length; i++) {
                 if (this.childSymbols[i].kind == symbolKind)
                     outList.push(this.childSymbols[i]);
-                this.childSymbols[i].collectChildSymbols(symbolKind, outList);
+                if (includeAllLevels)
+                    this.childSymbols[i].collectChildSymbols(symbolKind, includeAllLevels, outList);
             }
         }
     }
+
+    public collectChildSymbolsByName(findName: string, includeAllLevels: boolean, outList : AZSymbolInformation[]) {
+        if (this.childSymbols) {
+            for (let i=0; i<this.childSymbols.length; i++) {
+                if (this.childSymbols[i].name == findName)
+                    outList.push(this.childSymbols[i]);
+                if (includeAllLevels)
+                    this.childSymbols[i].collectChildSymbolsByName(findName, includeAllLevels, outList);
+            }
+        }
+    }
+
 
     public findFirstSymbolByKind(symbolKind: AZSymbolKind) : AZSymbolInformation | undefined {
         if (this.kind == symbolKind)