+++ src/symbollibraries/azSymbolKind.ts
@@ -7,7 +7,7 @@ export enum AZSymbolKind {
 
     CompilationUnit = 227,
 
-    //PropertyList = 228,
+    PropertyList = 228,
 
     //OptionValues = 229,
     //ObjectReference = 230,